connection_keeper: get rid of unused parts
Most of the responsibility was moved from ConnectionKeeper to
NodeConnectionPool. The former is still used for the control connection,
but some of its functionality became unused and was removed:

- Ability to set keyspace
- Sending updated ShardInfo

Moreover, the `many_connections` test was removed because a very similar
version of the test was introduced in `connection_pool.rs`.